chore: update bundle size numbers for v2
