+++ README.md
@@ -1,5 +1,7 @@
 # aws-sdk-js-bundle-size-comparisons
 
+- `yarn` to install all dependencies
+- `cd packages/code`
 - `yarn build` to compile JavaScript to bundles
 - `yarn test` to test source code
 - Examine the size of bundles emitted by webpack
@@ -9,11 +11,11 @@
 - @aws-sdk/v2 (VERSION: aws-sdk@2.646.0)
 
   ```console
-          Asset     Size  Chunks             Chunk Names
-  createNote.js  306 KiB       0  [emitted]  createNote
-  deleteNote.js  306 KiB       1  [emitted]  deleteNote
-     getNote.js  306 KiB       2  [emitted]  getNote
-  updateNote.js  306 KiB       3  [emitted]  updateNote
+  $ ls -lh dist | awk '{print $6,$10}'
+  305K createNote.js
+  305K deleteNote.js
+  305K getNote.js
+  305K updateNote.js
   ```
 
 - @aws-sdk/v3-preview (VERSION: @aws-sdk/client-dynamodb-node@0.1.0-preview.9)
