+++ packages/v2/package.json
@@ -1,9 +1,10 @@
 {
   "name": "@aws-sdk/v2",
   "version": "0.1.0",
-  "main": "index.js",
+  "main": "./src/index.js",
   "license": "MIT",
   "dependencies": {
     "aws-sdk": "^2.646.0"
-  }
+  },
+  "sideEffects": false
 }
