+++ packages/v2/src/deleteItem.js
@@ -1,0 +1,6 @@
+import { getClient } from "./getClient";
+
+export const deleteItem = async params =>
+  getClient()
+    .deleteItem(params)
+    .promise();
