+++ packages/v2/src/getClient.js
@@ -1,0 +1,3 @@
+import DynamoDB from "aws-sdk/clients/dynamodb";
+
+export const getClient = () => new DynamoDB({});
