+++ packages/v2/src/getItem.js
@@ -1,0 +1,6 @@
+import { getClient } from "./getClient";
+
+export const getItem = async params =>
+  getClient()
+    .getItem(params)
+    .promise();
