+++ packages/v2/src/index.js
@@ -1,0 +1,4 @@
+export * from "./putItem";
+export * from "./deleteItem";
+export * from "./updateItem";
+export * from "./getItem";
