+++ packages/v2/src/putItem.js
@@ -1,0 +1,6 @@
+import { getClient } from "./getClient";
+
+export const putItem = async params =>
+  getClient()
+    .putItem(params)
+    .promise();
