+++ packages/v2/src/updateItem.js
@@ -1,0 +1,6 @@
+import { getClient } from "./getClient";
+
+export const updateItem = async params =>
+  getClient()
+    .updateItem(params)
+    .promise();
